"Postmortem" -> "Retro"
This is nitpicky, but probs important to familiarize students with the correct use of Agile terminology. What is meant be "postmortem" throughout this reading is really a "retro".  Retros are done periodically _during_ the project, include the whole team, produce action items for incremental improvement (vs. postmortem: at the end / by management / produce a doc) 
https://softwaredevtools.com/blog/post-mortem-vs-retrospectives/
https://aspetraining.com/resources/blog/agile-retrospectives-vs-post-mortems